makefile: make $(EXAMPLE_FILE) target phony
Cargo knows about the dependencies for building an example, so we
make the $(EXAMPLE_FILE) target PHONY so Make will always ask cargo
to see what need to be rebuilt.

This has the downside of always causing the binfile to be regenerated,
even if there were no source changes.  Since that is a fast operation,
I believe this is the way to go for now.