build: convert target to lowercase for building paths
When building the example project, the build would fail on
case sensitive file systems as the environment variables passed
to the build script looks like CARGO_FEATURE_LPC17XX.  This resulted
in an attempt to copy a file from a directory that does not exist.

This change coverts the target to lowercase, which matches all
MCUs we currently have support for in zinc.