svd: migrate to using cmsis_svd repository for the parser implementation
The cmsis_svd python package is part of a larger effort in order
to share CMSIS SVD definitons and parsers amongst varioius projects.
The package also includes SVDs, but we continue to use our  own
for the time being until it is clear that we won't need to make
customizations.